% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load image}
\usage{
load_image(path)
}
\arguments{
\item{path}{String. Path to file.}
}
\value{
Cimg object.
}
\description{
Load image
}
\examples{
path <- system.file("./images/medicine.jpg", package = "mser4r")
image <- load_image(path)

}
