% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mser.R
\name{mser}
\alias{mser}
\title{Maximally Stable Extremal Regions}
\usage{
mser(
  image,
  scale = -1,
  steps = 50,
  min_area = 10,
  max_area = 10000,
  polarity = "positive",
  max_psi = 0.2,
  merge = TRUE
)
}
\arguments{
\item{image}{Cimg object. Input image.}

\item{scale}{Numeric. Scale factor for image resizing. When scale = -1, the scale factor is automatically computed such that the image is ~90000 pixels (this is 300x300 for a square image). Larger images result in longer computation. Default: -1.}

\item{steps}{Numeric. Number of steps in the sequence from 0 to 255. This determines the intensity levels where the region extraction takes place.}

\item{min_area}{Numeric. Minimum area in pixels to be considered a region.}

\item{max_area}{Numeric. Maximum area in pixels to be considered a region.}

\item{polarity}{String. Use "positive" to extract homogeneous areas with darker boundaries, "negative" to extract homogeneous areas with brighter boundaries, or "both" for both "positive" and "negative". Default: "positive".}

\item{max_psi}{Numeric. Stability score to determine whether a region is stable or not. Recommended values go from 0.2 to 1. Default: 0.2.}

\item{merge}{Logical. If TRUE, all the regions are merged into one single region. Default: TRUE.}
}
\value{
Cimg or imlist object.
}
\description{
Maximally Stable Extremal Regions
}
\examples{
image <- load_example()
regions <- mser(image, max_area = 2000)
plot_regions(regions, type = "mask", image_bg = image)

}
