% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regions.R
\name{plot_regions}
\alias{plot_regions}
\title{Plot merged regions}
\usage{
plot_regions(regions, type = "solid", image_bg = NULL)
}
\arguments{
\item{regions}{Regions object from mser() when merge = TRUE.}

\item{type}{String. Possible values are "border", "solid" or "mask".}

\item{image_bg}{Cimg object. Original image.}
}
\value{
Plot.
}
\description{
Plot merged regions
}
\examples{
image <- load_example()
regions <- mser(image, max_area = 2000)
plot_regions(regions, type = "mask", image_bg = image)

}
